WndRankTab only uses 1 .h
