+++ Source/_Interface/WndRankCmn.h
@@ -27,3 +27,83 @@ public:
 	virtual	void OnInitialUpdate();
 };
 
+
+class CWndRankTabBest : public CWndRankTab {
+public:
+	CWndRankTabBest();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankTabUnited : public CWndRankTab {
+public:
+	CWndRankTabUnited();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankTabPenya : public CWndRankTab {
+public:
+	CWndRankTabPenya();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankGuild : public CWndNeuz {
+public:
+	CWndRankTabBest		m_WndRankTabBest;
+	CWndRankTabUnited	m_WndRankTabUnited;
+	CWndRankTabPenya	m_WndRankTabPenya;
+
+	virtual BOOL Initialize(CWndBase * pWndParent = NULL, DWORD nType = MB_OK);
+	virtual	void OnInitialUpdate();
+};
+
+
+class CWndRankInfoTabLevel : public CWndRankTab {
+public:
+	CWndRankInfoTabLevel();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankInfoTabPlayTime : public CWndRankTab {
+public:
+	CWndRankInfoTabPlayTime();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankInfo : public CWndNeuz {
+public:
+	CWndRankInfoTabLevel	m_WndRankInfoTabLevel;
+	CWndRankInfoTabPlayTime m_WndRankInfoTabPlayTime;
+
+	virtual BOOL Initialize(CWndBase * pWndParent = NULL, DWORD nType = MB_OK);
+	virtual	void OnInitialUpdate();
+};
+
+
+class CWndRankWarTabGiveUp : public CWndRankTab {
+public:
+	CWndRankWarTabGiveUp();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankWarTabLose : public CWndRankTab {
+public:
+	CWndRankWarTabLose();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankWarTabWin : public CWndRankTab {
+public:
+	CWndRankWarTabWin();
+	ValuesToPrint GetValuesToPrint(const CGuildRank::GUILD_RANKING & ranking) override;
+};
+
+class CWndRankWar : public CWndNeuz {
+public:
+	CWndRankWarTabGiveUp	m_WndRankWarTabGiveUp;
+	CWndRankWarTabLose		m_WndRankWarTabLose;
+	CWndRankWarTabWin		m_WndRankWarTabWin;
+
+	virtual BOOL Initialize(CWndBase * pWndParent = NULL, DWORD nType = MB_OK);
+	virtual	void OnInitialUpdate();
+};
+
